requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
disnake==2.8.0
dnspython==2.3.0
easy-pil==0.1.6
frozenlist==1.3.3
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mafic==2.0.1
memoization==0.4.0
multidict==6.0.4
openai==0.27.4
Pillow==8.4.0
psutil==5.9.4
pycparser==2.21
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
yarl==1.8.2
